spin_oci
Module client

Spin’s client for distributing applications via OCI registries

Structs

Client

Client for interacting with an OCI registry for Spin applications.

ClientOpts

Options for configuring a Client

Enums

InferPredefinedAnnotations

Controls whether predefined annotations are generated when pushing an application. If an explicit annotation has the same name as a predefined one, the explicit one takes precedence.

Constants

ARCHIVE_MEDIATYPE

Media type for a layer representing a compressed archive of one or more files used by a Spin application

DATA_MEDIATYPE

Media type for a layer representing a generic data file used by a Spin application

SPIN_APPLICATION_MEDIA_TYPE

Media type for a layer representing a locked Spin application configuration

Functions

unpack_archive_layer

Unpack contents of the provided archive layer, represented by bytes and its corresponding digest, into the provided cache. A temporary staging directory is created via tempfile::tempdir() to store the unpacked contents prior to writing to the cache.